Decompose crs_matrix_vector_multiply into crs_row_vector_multiply etc.
In sparse.c, implement another version of Compressed Row Storage
matrix-vector multiply in which the function is decomposed into
one that handles a single row, and another calling that for each row.
The purpose is for algorithms that mix a single-row multiply with
other operations, such as some versions of Jacobi iteration.

And it's proved correct in verif_sparse_byrows.v.